+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Changed desktop display missions count to 5
   * Show loading icon only when loading
   * Show condensed date when possible
+  * Removed locale selection
 
 1.0.0-RC3 (2012-03-08)